prevent underflow in parse_number on check for Hex, Octal, or Binary
pc_length = pc_temp.len(); ---> 0
ch        = pc_temp.str[pc_length - 1];

changes some length return types to size_t

+++ src/uncrustify_types.h
@@ -263,7 +263,7 @@ struct chunk_t
    }
 
 
-   int len()
+   size_t len()
    {
       return(str.size());
    }
